pyproject.toml
[project]
name = "xppm-tdqn"
version = "0.1.0"
description = "Explainable prescriptive process monitoring with TDQN and offline RL"
authors = [{ name = "Andrew", email = "andrew@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "pyyaml",
    "scikit-learn",
    "torch",
    "tqdm",
    "pydantic",
    "fastapi",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "black",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


